move cards reducer
