+++ fronts-client/src/actions/__tests__/Cards.spec.ts
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import clipboardReducer from '../../reducers/clipboardReducer';
 import groupsReducer from 'reducers/groupsReducer';
-import cardsReducer from '../../shared/reducers/cardsReducer';
+import cardsReducer from 'reducers/cardsReducer';
 import {
   createSelectGroupArticles,
   createSelectSupportingArticles,
